Changed address data type to int from int64

+++ internal/endpoint/http/requests/user/user_request.go
@@ -1,7 +1,7 @@
 package request
 
 type UserRequest struct {
-	Address   int64  `json:"address"`
+	Address   int    `json:"address"`
 	FirstName string `json:"firstName"`
 	LastName  string `json:"lastName"`
 }
